CONTRIBUTING.md: change table column order for staging
The graph right above the table has the order as `master` ⇒ `staging-next` ⇒ `staging`.
This corresponds with the workflow, i.e. automatic merges happen in that direction and the manual merging is in reverse direction of that.
As a first time reader of this document the table was very confusing due to the disparity of that.
With this change the table should read more fluently for people not familiar with the workflow since the table follows the step-by-step flow of commits.

Signed-off-by: benaryorg <binary@benary.org>