emacs.pkgs.seq: stop shadowing it
magit requires[1] seq 2.24.  seq from GNU Elpa satisfies that.
However, it is shadowed by the Emacs builtin one to workaround an old
bug[2] and the version of the builtin seq in Emacs 28 is only 2.23.
So magit is broken for Emacs 28 which is the default one in NixOS
23.11 and available in the unstable branch.

This patch fixes magit by stopping shadowing seq from GNU Elpa since
that old bug[2] is not relevant now.

Fixes https://github.com/NixOS/nixpkgs/issues/272019.

[1]: https://github.com/magit/magit/blob/f4ff817cb2a48f0f7887050c3be469c03a059567/lisp/magit.el#L27
[2]: https://github.com/NixOS/nixpkgs/pull/74936