pythonPackages.crate: skip broken tests
Failing since Pandas update, see [Hydra build][1]. This is sort of fixed in
upstream (see [commit][2]), but is unreleased and also requires adding
pueblo (not in nixpkgs) and its dependents (at least one not in
nixpkgs needed for Python <3.11 only). Punt dealing with that till
updating this package, and make it complain at that time.

[1]: https://hydra.nixos.org/build/251074103
[2]: https://github.com/crate/crate-python/commit/db7ba4d0e1f4f4087739a8f9ebe1d71946333979