mmctl: 9.2.2 → 9.5.1, use override on mattermost
This makes the mmctl package be an override on the mattermost package,
as these were both moved by upstream into the same monorepo and use the
same build system. Nevertheless, it makes sense to keep them in separate
packages, as the mmctl utility is usually used on a different machine
than the mattermost server.

The meta information is retained as it was before (via an override).