nixops_unstable -> nixops_unstable_minimal.withPlugins
Providing the whole set by default is not feasible, and anything
smaller than that would be too arbitrary.

The aliases.nix error message puts users on the right path to get
exactly the plugins they need.

nixops_unstable_full probably won't be in a buildable state, so
we can't recommend it. It may be useful for CI.