grafana-agent: add `passthru.updateScript`
`nix-update` has support for `fetchYarnDeps` via the attribute name
`offlineCache`.

Our yarn FOD is called `frontendYarnOfflineCache`, which I find more
descriptive and would like to keep this way.

To work around this requirement, this adds an alias in `passthru`.

The addition of `passthru.updateScript` also allows `nixpkgs-update` to
update this package when the yarn FOD hash needs to be updated.

See the following error from the last `nixpkgs-update` run:
https://r.ryantm.com/log/grafana-agent/2024-02-27.log