Fix test failure
Specifically, the `nixosTest-example` test correctly flagged that the
change from 90679ea06c7f5ad3538d8c3940aba576865609bf broke the user's
expectation that the `pkgs` in `pkgs.nixosTest` would be reused as the
package set for the NixOS test system.

However, if we restore the original code removed by that commit:

```nix
nixpkgs.pkgs = lib.mkDefault pkgs;
```

… then that won't work because then we can't override `nixpkgs.system`
and reinstantiate that package set.  The whole *point* of this set of
changes is that we DON'T want to use the user's `pkgs` set as is
because then the wrong system is used for the NixOS test (on macOS).
Specifically, on Apple Silicon the `pkgs` in `pkgs.nixosTest` is going
to be instantiated with `system = aarch64-darwin;` but we want the
`nixpkgs.pkgs` that the NixOS machine uses to be that same package set
reinstantiated with `system = aarch64-linux;`.

Then I realized we can strike a balance by doing this instead:

```nix
nixpkgs.overlays = lib.mkDefault pkgs.overlays;
nixpkgs.config = lib.mkDefault pkgs.config;
```

… then we can get something which splits the difference: we preserve
the user's original package overrides but now we can reinstantiate
`nixpkgs.pkgs` with the new (correct) `nixpkgs.system`.

This isn't a foolproof solution, though, because `pkgs.overlays` and
`pkgs.config` do not necessarily contain all the information to
recreate the original `pkgs` that the user supplied.  As a concrete
example, if the user were to extend the package set "artificially" like
this:

```nix
(pkgs // { foo = …; }).nixosTest …
```

… then that `pkgs.foo` attribute would no longer be passed along to
`nixpgs.pkgs`.  However, in practice I believe most users are going to
be overriding `pkgs` using either the `config.packageOverrides` or
`overlays` arguments to the Nixpkgs function, so I think this is a
reasonable compromise.  Moreover, the benefit of this compromise is
that the NixOS test framework is much more "fire and forget" (we don't
need to instruct the users to instantiate Nixpkgs themselves twice for
the host and guest architecture).