gamja: init at 1.0.0-beta.9
gamja is an IRC web client.
Project url: https://git.sr.ht/~emersion/gamja

This is almost a direct copy of #243815 with an updated version. The
original PR seems to have stalled, although it was quite complete. I
have been using it for a couple of months, let's get this upstream.

Closes #243815

Co-authored-by: apfelkuchen06 <apfelkuchen@hrnz.li>