build-support/php: remove the dev packages information when they are not needed
Composer checks that all the requirements of the dev packages are met when
we run our final `composer update --lock` (even with `--no-dev`). This is
an issue because the dev packages can have additional requirements that are
not needed to run the production code.

Fixes the build of `php81Packages.php-codesniffer`