llvmPackages_16.mlir: init
The Nixpkgs release code generates a list of attributes on
x86_64-linux, then tries to evaluate them on all platforms.  This
meant that Darwin evals broke when llvmPackages was bumped to 17 on
Linux, because the list of attributes is evaluated for Linux, finds an
llvmPackages.mlir attribute, then gets an evaluation error when it
tries to evalute that on Darwin.  The easiest fix is to just make sure
an llvmPackages.mlir attribute exists on Darwin too.