libmicrohttpd: Refactor to inherit
The `pkgs/by-name` check currently fails on the staging-next merge into
master (https://github.com/NixOS/nixpkgs/actions/runs/8110296543/job/22167262748)
because it contains the changes of https://github.com/NixOS/nixpkgs/pull/278920, which introduced a "new" package using `callPackage` but not using `pkgs/by-name`.

This was never noticed in that PR because CI last ran almosts 2 months ago, which is before the check for new packages was introduced.

This wouldn't be a problem normally, it's only become a problem because
of the staging-next merge into master, which effectively PRs the same
change again (and the `pkgs/by-name` check doesn't try to distinguish
between branches).

The fix for this is a bit special because it's not actually a "new"
package, but rather just a new version of an existing package. The
`pkgs/by-name` check can't distinguish between such cases though.

So instead we make sure that the `pkgs/by-name` check doesn't think of
it as a package using `callPackage` by using the recommendation for
multi-versioned packages from
https://github.com/NixOS/nixpkgs/pull/292214.