grafana: explicitly declare subPackages
The build itself is pretty quick now:

    buildPhase completed in 2 minutes 46 seconds

in contrast to

    buildPhase completed in 5 minutes 22 seconds

before on the same machine (Hetzner AX41 NVMe with
AMD Ryzen 5 3600 6-Core Processor and 64GiB RAM).

Downside of that is that no tests are now executed, but I'm inclined to
make that sacrifice for now. Especially considering that a fix for that
is on the horizon with #284568.