formats: introduce jsonWithRuntimeSubstitution
This format allows an elegant and robust definition of settings which
are to be read from files or systemd credentials at runtime. Previous
solutions based on envsubst or similar brittle tooling had escaping
problems and no awareness of structured data.

TODO: there are some bike sheds to be painted regarding naming and
where in `lib` the utility functions belong.

Co-Authored-By: Maximilian Bosch <maximilian@mbosch.me>