homepage-dashboard: patch to enable control of log targets
This piece of software is generally used (and documented upstream)
in containers. It is common practice for such applications to log
both to stdout, and to a file which can be volume-mounted in the
container.

Because the intention here is to start with systemd, it makes less
sense to have both, so this added patch stops the application
from trying to create a log directory alongside the config directory.