skawarePackages: add manpages to their respective packages
For a while now, we’ve had these separate derivations for the skaware
manpages.

That was fine in the beginning, because it was not entirely clear
whether the manpage project would be long-lived. Given that the
collection of third-party manpages is now extensive and updated
regularly, plus it is sanctioned by skarnet, we can remove this
additional hurdle to using skaware.

The buildManPage structure is kept, instead of building them in a
separate `default.nix`, we add a field `manpages` to
`skawarePackages.buildPackage`, which adds the `"man"` output and
copies everything from the manpages output.

For backwards-compat, the manpage derivation is exposed in the
`passthru` and referenced by the `*-man-pages` attributes.

~~~

The `with skawarePackages;` scope is removed from all packages, and
used explicitly for all functions, while packages get added to the
package import header.