zoom-us: construct derivation on unsupported systems
It seems to be better to construct the derivation even on unsupported systems.
For example, OfBorg and nixpkgs CI automation are confused when a package derivation can not be queried for the package `name`.

`mkDerivation` will show an appropriate error message if it is evaluated on an unsupported platform.
The list of supported platforms is already specified in `meta.platforms`, which provides an excellent error message.

Co-authored-by: Philip Taron <philip.taron@gmail.com>
(cherry picked from commit 4dc2cc7805f1eb5db8b00696a9c9b09d047da1ec)