blender: rename local python{,Packages} -> python3{,Packages}
This makes them match what the fn param names *would* be, making it
easier to refactor e.g. to `{ python3Packages, ... }: mkDerivation …` in
the future if desired. It also potentially reduces first-glance
confusion that it might be Python 2 in-use.

Leave the passthru names as-is to preserve backwards-compatibility, but
also as there appears to be some precedent for naming them like this.