nixos-option: update to nix 2.18
This utility has horrendous code quality, partially due to nix's fault,
and really desperately needs to be rewritten, hopefully to not use the
nix C++ API. Either use the Python bindings to the Nix C API when the
Nix team merges the C API, or just write a clever bit of Nix script to
extract the data from a config, turn it into json, then process it into
good output in a CLI wrapper.

It currently does not support flakes and has many other issues.
(in my testing it did not accept -I nixos-path, it exploded due to
having plugins in /etc/nix/nix.conf with no way to override that, among
other breakage that simply does not exist in the Nix CLI).