postgresql: enable parallel building on darwin
This was disabled to fix #51093. The related discussion upstream revealed that this
has been fixed in [1] for v12+. Since v12 is our oldest supported version right now,
we should be able to safely enable parallel building on darwin again.

Furthermore, it seems that this kind of issue was also fixed in #51221 and #51408
for the general case.

[1]: https://github.com/postgres/postgres/commit/826eff57c4c23f77314ba7151d3dc506ce0fa24c