postgresql_12.pkgs.pg_repack: fix build
This failed because libxcrypt, which pg12 is linked against, was missing as a
build input. In general, this is an indication that this extension needs the
same build inputs that postgresql was built with - so we just pass those instead.

This makes similar build failures less likely in the future.