libpq: init at 16.2
libpq is just a name for the latest postgresql release currently in nixpkgs.
The idea is to use this for all packages depending on libpq, thus making
sure they always link against the latest version.

Currently everything links to the "postgresql" package, which is at v15,
not v16.

Since libpq is backwards-compatible, this should not be a problem.

Resolves #61580