nushellPlugins.regex: remove
The current version in nixpkgs is no longer compatible with the current
version of Nushell, and upstream is tracking the development branch of
Nushell, and is incompatible with the stable release of Nushell.

Background: https://github.com/fdncred/nu_plugin_regex/issues/3

This only made it into nixpkgs because there was a window in which the
Nushell plugin API in the released version and the version used by this
plugin were compatible.