nixos/postgresql: fix enableJIT = false
The enableJIT = true case was fixed in #221851 or
e2fb65175228a992f196f3b1700a53e18602e7f6 respectively.

However this did not take the case into consideration, when doing this:

    services.postgresql = {
      enable = true;
      enableJIT = false;
      package = pkgs.postgresql_15_jit;
    };

If enableJIT is treated as the source of truth, then this should indeed
cause JIT to be disabled, which this commit does.