i3: add more dependencies used by tests
Some tests in the i3 testsuite are silently skipped if the dependencies
required to run those tests are missing.  The problematic dependencies
turned out to be `xdotool`, `xorg.setxkbmap`, `xorg.xrandr`, `which`
(the latter is used to verify the presence of some tools).  Add those
dependencies to `buildInputs`, so that all tests would be run.

Ideally such missing dependencies should be detected as test failures,
but it's hard to implement, because missing tools result just in `skip`
marks in the test log, and there are some tests which are skipped for
other reasons (e.g., tests which were written before the expected
behavior is actually implemented in i3).