doc: update `toPythonApplication` section according to RFC 140
The advent of `pkgs/by-name` hierarchy created an interesting method:
instead of cluttering `all-packages.nix` with more lines and more merge
conflicts, the creation of `toPythonApplication` expressions can be made at
`pkgs/by-name`.

This method should be the default for new packages, while the old one should be
phased out.