firefox*: fix collisions due to identical `binaryName`
All `nixpkgs` Firefox builds are built with the same `binaryName`, and
as a result these packages collide when installing the unwrapped
packages, or even in the case of the wrapped packages, although the
`nameSuffix` is set differently for `firefox-beta`,
`firefox-devedition`, and `firefox-esr-115`, inside of the wrapper a
`".${binaryName}-old"` file is created, and _these_ still collide with
one another.

In NixOS, this is not an issue because the call to `buildEnv` allows
collisions, but in the case of Home Manager, its call to `buildEnv` does
not, and so one cannot install multiple versions of these packages at
the same time.

The `-bin` variants already have different binary names, as Mozilla
builds them with different binary names. I have adjusted the `nixpkgs`
builds of Firefox to use different binary names.