firefox*: fix `nativeMessagingHosts`
`wrapFirefox` calls `lib.getBin` on each of the packages in
`nativeMessagingHosts`. I checked `plasma-browser-integration` and
`tridactyl-native`, they both don't have a `bin` output defined, nor any
other specific output, which IIRC means that everything that is built is
put in the default `out` output that a derivation normally gives you.
The native messenger files actually live in the `lib` directory.

What this means is that the wrapper script cannot find the native
messaging files in the packages because it's looking in the wrong
output.

By instead looking in the default output of the derivations, the native
messaging JSON files are found correctly.

Fixes https://github.com/NixOS/nixpkgs/issues/281710.