pulsar: 1.109.0 -> 1.114.0
Release notes: https://github.com/pulsar-edit/pulsar/releases/tag/v1.114.0

pulsar has been marked insecure so this change would not be verified by
ofborg. Still, it is very actively developed and quite enjoyable to use.
As an atom veteran, I would like to help keep it usable; thus adding
myself to the maintainers.

A patchelf fixup for `ctags-linux` is adjusted according to the new
location of the binary.

Co-authored-by: Peder Bergebakken Sundt <pbsds@hotmail.com>