WIP: init stdenv.makeDerivationArgument
This is not complete, as described in the comment.

You can now do:

    nix-repl> :b derivation (stdenv.makeDerivationArgument { name = "hi"; dontUnpack = true; buildPhase = "echo hi > $out"; dontInstall = true; })

or even:

    nix-repl> :b x // { type = "derivation"; }
    error: derivation name missing

    nix-repl> x.drvPath
    "/nix/store/zfnznggsbm48y2g130wg1h891gwbf6s4-hi.drv"

    nix-repl> :b x // { name = "okay"; type = "derivation"; }

    This derivation produced the following outputs:
      out -> /nix/store/szky5bcmil93q0jjqprzc52mia5nfgsc-hi

The latter is obviously NOT recommended, because it does not
produce a standard package attribute set. Point is, new code
could implement that logic without duplicating what's in
Nix's built-in derivation.nix file.
Among many other improvements.
 - For instance, no cmake/meson overhead where it's not needed.
 - Single fixpoint https://github.com/NixOS/nixpkgs/issues/273815