appium-inspector: init at 2024.3.1
Add a new package for Appium Inspector, a GUI inspector for mobile apps.

Appium Inspector is a powerful tool for inspecting and debugging mobile
applications using the Appium automation framework. This commit adds a
new derivation to package version 2024.3.1 of Appium Inspector for
Nixpkgs.

The derivation fetches the AppImage release from the official GitHub
repository and wraps it using  to create a Nix-compatible
package. The necessary desktop file and icon are extracted from the
AppImage and installed into the appropriate locations.

The package is currently only available for the  platform.