Update Folding@home client to 8.3.7
Depend re2, libevent during build sisnce 8.3.7 appears to link with
them. Depend on git because client wan’t to know what revision it
was built from.

Add zlib and expat to FHS environment because some workunits depend on
them and cause lots of rejections if those libraries are not
available. Unfortunately they don’t appear to be bundled as part
of work distribution.