llvmPackages_11.libcxx: symlink cxxabi libs into libcxx
present a unified libcxx / libcxxabi lib directory similar to LLVM12+
toolchains. This is required to for building darwin stdenvBootstrapTools which pins
LLVM11 causing it to bootstrap a llvm11 stdenv but the bootstrapping
code expects a unified libcxx / libcxxabi. Thus, builds which explicitly
link to -lc++abi break as the libcxxabi is not in the linker search
path. Symlinking libc++abi into libcxx/lib fixes this.